requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv==1.0.0

python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

stripe

redis==5.0.1
celery==5.3.6

reportlab

slowapi

email-validator

pydantic-settings
alembic
gunicorn==21.2.0